Refactor: Move traffic tests to new file
Created new file to contain tests related to traffic charges, and moved
existing tests to that file.
